\name{Boik.mtx}
\alias{Boik.mtx}
\docType{data}
\title{Multi-headed Machine Data}
\description{
 Performance of a multiple-headed machine used to fill bottles. Weights for six heads on five occasions were recorded. 
}
\usage{data(Boik.mtx)}
\source{
  Boik, RJ. (1993) A comparison of three invariant tests of additivity in two-way classifications with no replications, Computational Statistics & Data Analysis, 15:411-424.
}
\examples{
data(Boik.mtx)
Boik.mtx
}
\keyword{datasets}
