\name{graybill.mtx}
\alias{graybill.mtx}
\docType{data}
\title{Wheat Yields}
\description{
Wheat yields from four genotypes in randomized block design with 13 locations.
}
\usage{data(graybill.mtx)}
\value{\item{graybill.mtx}{Matrix of wheat yields, rows are locations, columns are genotypes}
}
\source{
  Graybill, FA. (1954) Variance Heterogeneity in a Randomized Block Design, Biometrics, 10:516-520.
}
\examples{
data(graybill.mtx)
graybill.mtx
}
\keyword{datasets}
